worker/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
supabase==2.4.0
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
